Testing: Move editorbot fixture to conftest.py
