+++ spyder/plugins/editor/widgets/tests/test_goto.py
@@ -17,7 +17,6 @@ import pytest
 
 # Local imports
 from spyder.config.base import running_in_ci
-from spyder.plugins.editor.widgets.tests.test_codeeditor import editorbot
 from spyder.utils.vcs import get_git_remotes
 
 # Constants
@@ -83,7 +82,7 @@ TEST_FILE_REL = 'conftest.py'
     )
 def test_goto_uri(qtbot, editorbot, mocker, params):
     """Test that the uri search is working correctly."""
-    _, code_editor = editorbot
+    code_editor = editorbot
     code_editor.show()
     mocker.patch.object(QDesktopServices, 'openUrl')
 
@@ -123,7 +122,7 @@ def test_goto_uri(qtbot, editorbot, mocker, params):
 
 def test_goto_uri_project_root_path(qtbot, editorbot, mocker, tmpdir):
     """Test that the uri search is working correctly."""
-    _, code_editor = editorbot
+    code_editor = editorbot
     code_editor.show()
     mock_project_dir = str(tmpdir)
     expected_output_path = os.path.join(mock_project_dir, "some-file.txt")
@@ -174,7 +173,7 @@ def test_goto_uri_message_box(qtbot, editorbot, mocker):
     Test that a message box is displayed when the shorthand issue notation is
     used (gh-123) indicating the user that the file is not under a repository
     """
-    _, code_editor = editorbot
+    code_editor = editorbot
     code_editor.filename = TEMPFILE_PATH
     code_editor._last_hover_pattern_key = 'issue'
 
@@ -198,7 +197,7 @@ def test_goto_uri_message_box(qtbot, editorbot, mocker):
 
 def test_pattern_highlight_regression(qtbot, editorbot):
     """Fix regression on spyder-ide/spyder#11376."""
-    _, code_editor = editorbot
+    code_editor = editorbot
     code_editor.show()
 
     # This was generating an infinite loop
